+++ src/common/utils/index.ts
@@ -46,7 +46,9 @@ export function git(options: any = { output: false }): SimpleGit {
     function logCommand() {
       if (first) {
         first = false;
-        uxLog(this, `[command] ${c.grey(command)} ${c.grey(gitArgs.join(" "))}`);
+        if (!(gitArgs && gitArgs[0] && gitArgs[0] === "branch" && gitArgs[1] && gitArgs[1] === "-v")) {
+          uxLog(this, `[command] ${c.grey(command)} ${c.grey(gitArgs.join(" "))}`);
+        }
       }
     }
   });
@@ -132,7 +134,7 @@ export async function promptInstanceUrl() {
         value: "https://test.salesforce.com",
       },
       {
-        title: "Other: Dev, Enterprise or DevHub org... (login.salesforce.com)",
+        title: "Other: Dev, Enterprise or DevHub org (login.salesforce.com)",
         description: "The org I want to connect is NOT a sandbox",
         value: "https://login.salesforce.com",
       },